+++ .semversioner/next-release/patch-20241213181544864279.json
@@ -1,0 +1,4 @@
+{
+  "type": "patch",
+  "description": "Move extractor code to co-locate with operations."
+}
