--- graphrag/index/graph/extractors/__init__.py
@@ -1,17 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""The Indexing Engine graph extractors package root."""
-
-from graphrag.index.graph.extractors.claims import ClaimExtractor
-from graphrag.index.graph.extractors.community_reports import (
-    CommunityReportsExtractor,
-)
-from graphrag.index.graph.extractors.graph import GraphExtractionResult, GraphExtractor
-
-__all__ = [
-    "ClaimExtractor",
-    "CommunityReportsExtractor",
-    "GraphExtractionResult",
-    "GraphExtractor",
-]
