--- graphrag/index/graph/extractors/claims/__init__.py
@@ -1,8 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""The Indexing Engine graph extractors claims package root."""
-
-from graphrag.index.graph.extractors.claims.claim_extractor import ClaimExtractor
-
-__all__ = ["ClaimExtractor"]
