--- graphrag/index/graph/extractors/community_reports/__init__.py
@@ -1,24 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""The Indexing Engine community reports package root."""
-
-import graphrag.index.graph.extractors.community_reports.schemas as schemas
-from graphrag.index.graph.extractors.community_reports.build_mixed_context import (
-    build_mixed_context,
-)
-from graphrag.index.graph.extractors.community_reports.community_reports_extractor import (
-    CommunityReportsExtractor,
-)
-from graphrag.index.graph.extractors.community_reports.prep_community_report_context import (
-    prep_community_report_context,
-)
-from graphrag.index.graph.extractors.community_reports.sort_context import sort_context
-
-__all__ = [
-    "CommunityReportsExtractor",
-    "build_mixed_context",
-    "prep_community_report_context",
-    "schemas",
-    "sort_context",
-]
