--- graphrag/index/graph/extractors/graph/__init__.py
@@ -1,16 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""The Indexing Engine unipartite graph package root."""
-
-from graphrag.index.graph.extractors.graph.graph_extractor import (
-    DEFAULT_ENTITY_TYPES,
-    GraphExtractionResult,
-    GraphExtractor,
-)
-
-__all__ = [
-    "DEFAULT_ENTITY_TYPES",
-    "GraphExtractionResult",
-    "GraphExtractor",
-]
