--- graphrag/index/graph/extractors/summarize/__init__.py
@@ -1,11 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""The Indexing Engine unipartite graph package root."""
-
-from graphrag.index.graph.extractors.summarize.description_summary_extractor import (
-    SummarizationResult,
-    SummarizeExtractor,
-)
-
-__all__ = ["SummarizationResult", "SummarizeExtractor"]
