--- graphrag/index/graph/visualization/__init__.py
@@ -1,17 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""The Indexing Engine graph visualization package root."""
-
-from graphrag.index.graph.visualization.compute_umap_positions import (
-    compute_umap_positions,
-    get_zero_positions,
-)
-from graphrag.index.graph.visualization.typing import GraphLayout, NodePosition
-
-__all__ = [
-    "GraphLayout",
-    "NodePosition",
-    "compute_umap_positions",
-    "get_zero_positions",
-]
