+++ graphrag/index/operations/embed_graph/typing.py
@@ -4,7 +4,19 @@
 """A module containing different lists and dictionaries."""
 
 # Use this for now instead of a wrapper
+from enum import Enum
 from typing import Any
+
+
+class EmbedGraphStrategyType(str, Enum):
+    """EmbedGraphStrategyType class definition."""
+
+    node2vec = "node2vec"
+
+    def __repr__(self):
+        """Get a string representation."""
+        return f'"{self.value}"'
+
 
 NodeList = list[str]
 EmbeddingList = list[Any]
