--- graphrag/index/operations/extract_entities/strategies/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""The Indexing Engine entities extraction strategies package root."""
