+++ graphrag/index/operations/layout_graph/__init__.py
@@ -2,9 +2,3 @@
 # Licensed under the MIT License
 
 """The Indexing Engine graph layout package root."""
-
-from graphrag.index.operations.layout_graph.layout_graph import (
-    layout_graph,
-)
-
-__all__ = ["layout_graph"]
