+++ graphrag/index/operations/layout_graph/typing.py
