+++ graphrag/index/operations/summarize_communities/community_reports_extractor/sort_context.py
@@ -4,7 +4,7 @@
 
 import pandas as pd
 
-import graphrag.index.graph.extractors.community_reports.schemas as schemas
+import graphrag.index.operations.summarize_communities.community_reports_extractor.schemas as schemas
 from graphrag.query.llm.text_utils import num_tokens
