+++ graphrag/index/operations/summarize_communities/summarize_communities.py
@@ -15,12 +15,14 @@ from datashaper import (
 )
 
 import graphrag.config.defaults as defaults
-import graphrag.index.graph.extractors.community_reports.schemas as schemas
+import graphrag.index.operations.summarize_communities.community_reports_extractor.schemas as schemas
 from graphrag.cache.pipeline_cache import PipelineCache
-from graphrag.index.graph.extractors.community_reports import (
+from graphrag.index.operations.summarize_communities.community_reports_extractor import (
     prep_community_report_context,
 )
-from graphrag.index.graph.extractors.community_reports.utils import get_levels
+from graphrag.index.operations.summarize_communities.community_reports_extractor.utils import (
+    get_levels,
+)
 from graphrag.index.operations.summarize_communities.typing import (
     CommunityReport,
     CommunityReportsStrategy,
